+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: fbarrett <fbarrett@student.42quebec>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 14:31:34 by fbarrett          #+#    #+#             */
-/*   Updated: 2024/01/14 17:39:39 by fbarrett         ###   ########.fr       */
+/*   Updated: 2024/01/15 12:13:20 by fbarrett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ void	parent_close(s_pipe *pipe)
 	free(pipe->child_list);
 }
 
-void	child_process(s_pipe *pipe, char **line)
+void	close_child(s_pipe *pipe)
 {
 	int ite;
 
@@ -100,6 +100,13 @@ void	child_process(s_pipe *pipe, char **line)
 			close(ite);
 		ite--;
 	}
+}
+
+void	child_process(s_pipe *pipe, char **line)
+{
+	int ite;
+
+	close_child(pipe);
 	ite = 0;
 	while (ft_strncmp("|", line[pipe->cmd_ptr + ite], 2))
 		ite++;
@@ -113,19 +120,8 @@ void	child_process(s_pipe *pipe, char **line)
 	pipe->cmd_args[ite] = 0;
 }
 
-int	run_cmds_pipe(char **line, char	**cmd_paths, char **envp, s_pipe *pipes)
+int	run_each_cmd(s_pipe *pipes, char **cmd_paths, char **envp, char **line)
 {
-	pipes->i = 0;
-	pipes->cmd_ptr = 0;
-	pipes->child_list = ft_calloc(pipes->pipes_nbr + 2, sizeof(int));
-	while (pipes->i < pipes->pipes_nbr)
-	{
-		pipe(pipes->fd);
-		if (pipes->i == 0)
-			pipes->min_fd = pipes->fd[0];
-		pipes->i++;	
-	}
-	pipes->max_fd = pipes->fd[1];
 	while (pipes->i >= 0)
 	{
 		if	((pipes->child = fork()) < 0)
@@ -144,30 +140,29 @@ int	run_cmds_pipe(char **line, char	**cmd_paths, char **envp, s_pipe *pipes)
 		free_all(pipes->cmd_args);
 		exit (1);
 	}
-	parent_close(pipes);
 	return (0);
 }
 
-void	seek_all_cmds(char ***cmd_paths, char **line_args, char **envp)
+int	run_cmds_pipe(char **line, char	**cmd_paths, char **envp, s_pipe *pipes)
 {
-	int i;
-	int args;
-
-	args = 1;
-	i = 1;
-	(*cmd_paths)[0] = seek_cmd(line_args[0], envp);
-	while (line_args[i] && (*cmd_paths)[0])
+	pipes->i = 0;
+	pipes->cmd_ptr = 0;
+	pipes->child_list = ft_calloc(pipes->pipes_nbr + 2, sizeof(int));
+	while (pipes->i < pipes->pipes_nbr)
 	{
-		if (!ft_strncmp("|", line_args[i - 1], 2))
-		{
-			(*cmd_paths)[args] = seek_cmd(line_args[i], envp);
-			args++;
-			if (!(*cmd_paths)[args - 1])
-				break ;
-		}
-		i++;
+		pipe(pipes->fd);
+		if (pipes->i == 0)
+			pipes->min_fd = pipes->fd[0];
+		pipes->i++;	
 	}
-	(*cmd_paths)[args] = 0;
+	pipes->max_fd = pipes->fd[1];
+	if (run_each_cmd(pipes, cmd_paths, envp, line))
+	{
+		parent_close(pipes);
+		return (1);
+	}
+	parent_close(pipes);
+	return (0);
 }
 
 int	check_cmds(s_pipe *pipe, char **cmd_paths)
@@ -184,16 +179,45 @@ int	check_cmds(s_pipe *pipe, char **cmd_paths)
 	return (0);
 }
 
+void	free_moi_ca(char *buff, char **cmd_paths, char **line_args)
+{
+	free_all(line_args);
+	free_all(cmd_paths);
+	free(buff);
+}
+
+int	exec_line(s_pipe *pipe, char **line_args, char **envp, char *buff)
+{
+	char	**cmd_paths;
+
+	pipe->pipes_nbr = seek_pipe(line_args);
+	if (pipe->pipes_nbr == 0)
+	{
+		cmd_paths = ft_calloc(2, sizeof(char *));
+		cmd_paths[0] = seek_cmd(line_args[0], envp);
+		cmd_paths[1] = 0;
+		if (cmd_paths[0])
+			run_single_cmd(line_args, cmd_paths[0], envp, pipe);
+	}
+	else
+	{
+		cmd_paths = ft_calloc(2 + pipe->pipes_nbr, sizeof(char *));
+		seek_all_cmds(&cmd_paths, line_args, envp);
+		if (!check_cmds(pipe, cmd_paths))
+			run_cmds_pipe(line_args, cmd_paths, envp, pipe);
+	}
+	free_moi_ca(buff, cmd_paths, line_args);
+	return (0);
+}
+
 int main(int argc, char	**argv, char **envp)
 {
 	char	*buff;
 	char	**line_args;
-	char	**cmd_paths;
 	s_pipe	*pipe;
 
 	(void)argc;
 	(void)argv;
-	(void)envp;
 	pipe = ft_calloc(1, sizeof(char *));
 	while (1)
 	{
@@ -216,25 +240,7 @@ int main(int argc, char	**argv, char **envp)
 			free(buff);
 			continue ;
 		}
-		pipe->pipes_nbr = seek_pipe(line_args);
-		if (pipe->pipes_nbr == 0)
-		{
-			cmd_paths = ft_calloc(2, sizeof(char *));
-			cmd_paths[0] = seek_cmd(line_args[0], envp);
-			cmd_paths[1] = 0;
-			if (cmd_paths[0])
-				run_single_cmd(line_args, cmd_paths[0], envp, pipe);
-		}
-		else
-		{
-			cmd_paths = ft_calloc(2 + pipe->pipes_nbr, sizeof(char *));
-			seek_all_cmds(&cmd_paths, line_args, envp);
-			if (!check_cmds(pipe, cmd_paths))
-				run_cmds_pipe(line_args, cmd_paths, envp, pipe);
-		}
-		free_all(line_args);
-		free_all(cmd_paths);
-		free(buff);
+		exec_line(pipe, line_args, envp, buff);
 	}
 	free(pipe);
 	// y faut rl_clear_history
